prom_telemetry-master/requirements.txt
requests==2.22.0
pyserial==3.5
